move DeploymentMode and RuntimeMode to stream.processor.core
