+++ components/org.wso2.carbon.das.jobmanager.core/src/main/java/org/wso2/carbon/das/jobmanager/core/internal/ServiceComponent.java
@@ -45,9 +45,9 @@ import org.wso2.carbon.das.jobmanager.core.impl.DistributionManagerServiceImpl;
 import org.wso2.carbon.das.jobmanager.core.impl.DistributionResourceServiceImpl;
 import org.wso2.carbon.das.jobmanager.core.impl.RDBMSServiceImpl;
 import org.wso2.carbon.das.jobmanager.core.model.ResourceMapping;
-import org.wso2.carbon.das.jobmanager.core.util.DeploymentMode;
+import org.wso2.carbon.stream.processor.core.util.DeploymentMode;
 import org.wso2.carbon.das.jobmanager.core.util.DistributedConstants;
-import org.wso2.carbon.das.jobmanager.core.util.RuntimeMode;
+import org.wso2.carbon.stream.processor.core.util.RuntimeMode;
 import org.wso2.carbon.datasource.core.api.DataSourceService;
 import org.wso2.carbon.stream.processor.core.distribution.DistributionService;
 import org.wso2.carbon.utils.Utils;
