+++ components/org.wso2.carbon.stream.processor.core/src/main/java/org/wso2/carbon/stream/processor/core/util/RuntimeMode.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.das.jobmanager.core.util;
+package org.wso2.carbon.stream.processor.core.util;
 
 public enum RuntimeMode {
     MANAGER, RESOURCE
